requirements.txt
numpy
psutil
Pillow
pygame-ce>=2.5.5,<3.0
pygame-gui==0.6.14
matplotlib